Pushing for grading
experienced an issue where locally the background images were set by url('images/filename.type').  Atom Live Server and Chrome were then not showing the images because the rendered url was showing as url('css/images/filename.type').  Had to edit the url for local development to url('../images/filename.type') so could work on responsiveness and then removed it for the push.